Move search to its own component
